Get all unprocessed screenshots
This change enables us to get all unprocessed screenshots without forcing the
user to limit their daily screenshots. Previously we'd only get the top 12,
meaning if the user took more than 12 screenshots between invocations, they
would not be processed.

This helps resolve the intial 'seed' problem too where a user might want to
post every screenshot they've taken so far. This solves that problem completely,
as the tool will now process all screenshots until it gets to the end.

There was an issue with Playwright where if we tried to pull too many screenshots
at the same time that it would timeout and fail. Not sure whether this was a local
issue or Steam putting us on the naughty step, but limiting the number of concurrent
processes resolves the issue. Haven't done this for Tinify yet - will look into that
after as it might have the same issue.